requirements.txt
# Telegram Bot Framework
python-telegram-bot[job-queue]==20.7

# Binance API
python-binance==1.0.19

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Async HTTP Client
aiohttp==3.9.1

# Task Scheduling
apscheduler==3.10.4

# Testing (optional for MVP)
pytest==7.4.3
pytest-asyncio==0.21.1
